+++ src/main/java/ai/classlynk/app/ClassLynkApplication.java
@@ -4,32 +4,30 @@ import ai.classlynk.data_access.APIDataAccessObject;
 import ai.classlynk.data_access.FirebaseDataAccessObject;
 import ai.classlynk.entity.SClass;
 import ai.classlynk.entity.Timetable;
-import ai.classlynk.interface_adapter.Login.LoginController;
-import ai.classlynk.interface_adapter.Login.LoginPresenter;
-import ai.classlynk.interface_adapter.Login.LoginViewModel;
+import ai.classlynk.interface_adapter.login.LoginController;
+import ai.classlynk.interface_adapter.login.LoginPresenter;
+import ai.classlynk.interface_adapter.login.LoginViewModel;
 import ai.classlynk.interface_adapter.MenuSwitchingController;
-import ai.classlynk.interface_adapter.Register.RegisterController;
-import ai.classlynk.interface_adapter.Register.RegisterPresenter;
-import ai.classlynk.interface_adapter.Register.RegisterViewModel;
-import ai.classlynk.interface_adapter.ViewCourse.ViewCourseController;
-import ai.classlynk.interface_adapter.ViewCourse.ViewCoursePresenter;
-import ai.classlynk.interface_adapter.ViewCourse.ViewCourseViewModel;
+import ai.classlynk.interface_adapter.register.RegisterController;
+import ai.classlynk.interface_adapter.register.RegisterPresenter;
+import ai.classlynk.interface_adapter.register.RegisterViewModel;
+import ai.classlynk.interface_adapter.view_course.ViewCourseController;
+import ai.classlynk.interface_adapter.view_course.ViewCoursePresenter;
+import ai.classlynk.interface_adapter.view_course.ViewCourseViewModel;
 import ai.classlynk.interface_adapter.ViewManagerModel;
-import ai.classlynk.interface_adapter.addToCart.AddToCartController;
-import ai.classlynk.interface_adapter.addToCart.AddToCartPresenter;
-import ai.classlynk.interface_adapter.addToCart.AddToCartViewModel;
+import ai.classlynk.interface_adapter.add_to_cart.AddToCartController;
+import ai.classlynk.interface_adapter.add_to_cart.AddToCartPresenter;
+import ai.classlynk.interface_adapter.add_to_cart.AddToCartViewModel;
 import ai.classlynk.interface_adapter.generate_timetable.GenerateTimetableController;
 import ai.classlynk.interface_adapter.save_view_timetables.SaveViewTimetableController;
 import ai.classlynk.interface_adapter.save_view_timetables.SaveViewTimetablePresenter;
 import ai.classlynk.interface_adapter.save_view_timetables.SaveViewTimetableViewModel;
 import ai.classlynk.interface_adapter.static_maps.*;
-import ai.classlynk.use_case.AddToCart.AddToCartInteractor;
-import ai.classlynk.use_case.ViewCourse.ViewCourseInteractor;
+import ai.classlynk.use_case.add_to_cart.AddToCartInteractor;
+import ai.classlynk.use_case.view_course.ViewCourseInteractor;
 import ai.classlynk.use_case.generate_timetable.GenerateTimetableInteractor;
 import ai.classlynk.use_case.save_view_timetables.SaveViewTimetableInteractor;
-import ai.classlynk.use_case.static_maps.MapsInputBoundary;
 import ai.classlynk.use_case.static_maps.MapsInteractor;
-import ai.classlynk.use_case.static_maps.MapsOutputBoundary;
 import ai.classlynk.use_case.user_auth.login.LoginInteractor;
 import ai.classlynk.use_case.user_auth.register.RegisterInteractor;
 import ai.classlynk.view.*;
