+++ src/main/java/ai/classlynk/data_access/FirebaseDataAccessObject.java
@@ -3,8 +3,8 @@ package ai.classlynk.data_access;
 import ai.classlynk.entity.Course;
 import ai.classlynk.entity.Timetable;
 import ai.classlynk.entity.User;
-import ai.classlynk.use_case.AddToCart.AddToCartDataAccessInterface;
-import ai.classlynk.use_case.ViewCourse.ViewCourseDataAccessInterface;
+import ai.classlynk.use_case.add_to_cart.AddToCartDataAccessInterface;
+import ai.classlynk.use_case.view_course.ViewCourseDataAccessInterface;
 import ai.classlynk.use_case.save_view_timetables.SaveViewTimetablesDataAccessInterface;
 import ai.classlynk.use_case.user_auth.login.LoginDataAccessInterface;
 import ai.classlynk.use_case.user_auth.register.RegisterDataAccessInterface;
