+++ src/main/java/ai/classlynk/interface_adapter/login/LoginPresenter.java
@@ -1,4 +1,4 @@
-package ai.classlynk.interface_adapter.Login;
+package ai.classlynk.interface_adapter.login;
 
 import ai.classlynk.entity.User;
 import ai.classlynk.interface_adapter.ViewManagerModel;
