+++ src/main/java/ai/classlynk/interface_adapter/register/RegisterViewModel.java
@@ -1,4 +1,4 @@
-package ai.classlynk.interface_adapter.Register;
+package ai.classlynk.interface_adapter.register;
 
 import ai.classlynk.interface_adapter.ViewModel;
