+++ src/main/java/ai/classlynk/use_case/add_to_cart/AddToCartDataAccessInterface.java
@@ -1,10 +1,6 @@
-package ai.classlynk.use_case.AddToCart;
+package ai.classlynk.use_case.add_to_cart;
 
-import ai.classlynk.entity.Course;
-import ai.classlynk.entity.SClass;
 import ai.classlynk.entity.User;
-
-import java.util.List;
 
 public interface AddToCartDataAccessInterface {
     public User getUser(String username);
