+++ src/main/java/ai/classlynk/use_case/add_to_cart/AddToCartInputBoundary.java
@@ -1,4 +1,4 @@
-package ai.classlynk.use_case.AddToCart;
+package ai.classlynk.use_case.add_to_cart;
 
 public interface AddToCartInputBoundary {
     void addToCart(AddToCartInputData addToCartInputData);
