+++ src/main/java/ai/classlynk/use_case/add_to_cart/AddToCartInputData.java
@@ -1,8 +1,6 @@
-package ai.classlynk.use_case.AddToCart;
+package ai.classlynk.use_case.add_to_cart;
 
 import ai.classlynk.entity.Course;
-import ai.classlynk.entity.SClass;
-import ai.classlynk.entity.ClassBundle;
 
 public class AddToCartInputData {
     private Course course;
