+++ src/main/java/ai/classlynk/use_case/add_to_cart/AddToCartInteractor.java
@@ -1,4 +1,4 @@
-package ai.classlynk.use_case.AddToCart;
+package ai.classlynk.use_case.add_to_cart;
 
 import ai.classlynk.entity.Course;
 import ai.classlynk.entity.User;
