+++ src/main/java/ai/classlynk/use_case/add_to_cart/AddToCartOutputBoundary.java
@@ -1,4 +1,4 @@
-package ai.classlynk.use_case.AddToCart;
+package ai.classlynk.use_case.add_to_cart;
 
 public interface AddToCartOutputBoundary {
     void presentResponse(AddToCartOutputData outputData);
