+++ src/main/java/ai/classlynk/use_case/view_course/ViewCourseOutputBoundary.java
@@ -1,4 +1,4 @@
-package ai.classlynk.use_case.ViewCourse;
+package ai.classlynk.use_case.view_course;
 
 public interface ViewCourseOutputBoundary {
     void presentResponse(ViewCourseOutputData outputData);
