+++ src/main/java/ai/classlynk/use_case/view_course/ViewCourseOutputData.java
@@ -1,4 +1,4 @@
-package ai.classlynk.use_case.ViewCourse;
+package ai.classlynk.use_case.view_course;
 
 import ai.classlynk.entity.Course;
