+++ src/main/java/ai/classlynk/view/RegisterView.java
@@ -1,10 +1,10 @@
 package ai.classlynk.view;
 
-import ai.classlynk.interface_adapter.Login.LoginState;
-import ai.classlynk.interface_adapter.Login.LoginViewModel;
-import ai.classlynk.interface_adapter.Register.RegisterController;
-import ai.classlynk.interface_adapter.Register.RegisterState;
-import ai.classlynk.interface_adapter.Register.RegisterViewModel;
+import ai.classlynk.interface_adapter.login.LoginState;
+import ai.classlynk.interface_adapter.login.LoginViewModel;
+import ai.classlynk.interface_adapter.register.RegisterController;
+import ai.classlynk.interface_adapter.register.RegisterState;
+import ai.classlynk.interface_adapter.register.RegisterViewModel;
 import ai.classlynk.interface_adapter.ViewManagerModel;
 import com.google.firebase.auth.FirebaseAuthException;
